test(crs): move tests to test_postgresql_provider
