+++ tests/test_api.py
@@ -44,7 +44,7 @@ from pygeoapi.api import (
     API, APIRequest, FORMAT_TYPES, validate_bbox, validate_datetime,
     validate_subset, F_HTML, F_JSON, F_JSONLD, F_GZIP
 )
-from pygeoapi.util import yaml_load, geojson_to_geom, get_transform_from_crs
+from pygeoapi.util import yaml_load
 
 from .util import get_test_file_path, mock_request
 
@@ -925,78 +925,6 @@ def test_get_collection_items_crs(config, api_):
                 break
 
 
-def test_get_collection_items_postgresql_crs(pg_api_):
-    """Test the coordinates transformation implementation of
-    PostgreSQLProvider when using the crs parameter.
-    """
-    storage_crs = 'http://www.opengis.net/def/crs/EPSG/0/4326'
-    crs_32735 = 'http://www.opengis.net/def/crs/EPSG/0/32735'
-
-    # Without CRS query parameter -> no coordinates transformation
-    req = mock_request({'bbox': '29.0,-2.85,29.05,-2.8'})
-    rsp_headers, code, response = pg_api_.get_collection_items(
-        req, 'hot_osm_waterways',
-    )
-
-    assert code == HTTPStatus.OK
-
-    features_orig = json.loads(response)
-
-    # With CRS query parameter not resulting in coordinates transformation
-    # (i.e. 'crs' query parameter is the same as 'storage_crs')
-    req = mock_request({'crs': storage_crs, 'bbox': '29.0,-2.85,29.05,-2.8'})
-    rsp_headers, code, response = pg_api_.get_collection_items(
-        req, 'hot_osm_waterways',
-    )
-
-    assert code == HTTPStatus.OK
-
-    features_4326 = json.loads(response)
-
-    # With CRS query parameter resulting in coordinates transformation
-    req = mock_request({'crs': crs_32735, 'bbox': '29.0,-2.85,29.05,-2.8'})
-    rsp_headers, code, response = pg_api_.get_collection_items(
-        req, 'hot_osm_waterways',
-    )
-
-    assert code == HTTPStatus.OK
-
-    features_32735 = json.loads(response)
-
-    # Make sure that we compare the same features
-    assert len(features_orig) == len(features_4326) == len(features_32735)
-    assert (
-        sorted(features_orig['features'])
-        == sorted(features_4326['features'])
-        == sorted(features_32735['features'])
-    )
-
-    # Without 'crs' query parameter or with 'crs' set to 'storage_crs', the
-    # geometries of the returned features should be the same
-    for feat_orig in features_orig['features']:
-        id_ = feat_orig['id']
-        for feat_4326 in features_4326['features']:
-            if id_ == feat_4326['id']:
-
-                assert feat_orig['geometry'] == feat_4326['geometry']
-                break
-
-    transform_func = get_transform_from_crs(
-        pyproj.CRS.from_epsg(4326),
-        pyproj.CRS.from_epsg(32735),
-    )
-    # Check that the coordinates of returned features were transformed
-    for feat_orig in features_orig['features']:
-        id_ = feat_orig['id']
-        geom_orig = geojson_to_geom(feat_orig['geometry'])
-        for feat_32735 in features_32735['features']:
-            if id_ == feat_32735['id']:
-                geom_32735 = geojson_to_geom(feat_32735['geometry'])
-
-                assert geom_32735.equals_exact(transform_func(geom_orig), 1)
-                break
-
-
 def test_get_collection_items_json_ld(config, api_):
     req = mock_request({
         'f': 'jsonld',
@@ -1123,71 +1051,6 @@ def test_get_collection_item_crs(config, api_):
     loc_4258 = Point(*feature_4258['geometry']['coordinates'])
 
     assert loc_4258.equals_exact(loc_transf, 1e-5)
-
-
-def test_get_collection_item_postgresql_crs(pg_api_):
-    """Test the coordinates transformation implementation of
-    PostgreSQLProvider when using the crs parameter.
-    """
-    storage_crs = 'http://www.opengis.net/def/crs/EPSG/0/4326'
-    crs_32735 = 'http://www.opengis.net/def/crs/EPSG/0/32735'
-    # List of feature IDs located in UTM zone 35S
-    fid_list = [
-        '439338397',
-        '198190856',
-        '93063941',
-        '586449587',
-        '80827793',
-        '587350255',
-        '586994284',
-        '587960337',
-        '586449586',
-        '422440125',
-    ]
-    for fid in fid_list:
-        # Without CRS query parameter -> no coordinates transformation
-        req = mock_request()
-        rsp_headers, code, response = pg_api_.get_collection_items(
-            req, 'hot_osm_waterways', fid,
-        )
-
-        assert code == HTTPStatus.OK
-
-        feat_orig = json.loads(response)
-        geom_orig = geojson_to_geom(feat_orig['geometry'])
-
-        # With CRS query parameter not resulting in coordinates transformation
-        # (i.e. 'crs' query parameter is the same as 'storage_crs')
-        req = mock_request({'crs': storage_crs})
-        rsp_headers, code, response = pg_api_.get_collection_items(
-            req, 'hot_osm_waterways', fid,
-        )
-
-        assert code == HTTPStatus.OK
-
-        feat_4326 = json.loads(response)
-
-        # Without 'crs' query parameter or with 'crs' set to 'storage_crs', the
-        # geometries should be identical
-        assert feat_orig['geometry'] == feat_4326['geometry']
-
-        # With CRS query parameter resulting in coordinates transformation
-        req = mock_request({'crs': crs_32735})
-        rsp_headers, code, response = pg_api_.get_collection_items(
-            req, 'hot_osm_waterways', fid,
-        )
-
-        assert code == HTTPStatus.OK
-
-        feat_32735 = json.loads(response)
-        geom_32735 = geojson_to_geom(feat_32735['geometry'])
-
-        transform_func = get_transform_from_crs(
-            pyproj.CRS.from_epsg(4326),
-            pyproj.CRS.from_epsg(32735),
-        )
-        # Check that the coordinates of returned feature were transformed
-        assert geom_32735.equals_exact(transform_func(geom_orig), 1)
 
 
 def test_get_collection_item_json_ld(config, api_):
